mon_scanner/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
Jinja2==3.1.3
PyYAML==6.0.1
urllib3==2.2.0
